+++ public/index.html
@@ -122,7 +122,8 @@
 
 </main>
 
+<script src="js/components/leaderBoard.js"></script>
 <script src="js/modules/http.js"></script>
-<script src="js/application.js"></script>
+<script type="module" src="js/application.js"></script>
 </body>
 </html>
