+++ public/js/application.js
@@ -1,6 +1,6 @@
 'use strict';
 
-
+import FormError from "./components/validation.js";
 const httpModule = new window.HttpModule();
 
 const application = document.getElementById('application');
@@ -21,6 +21,9 @@ loginSection.style.display = 'none';
 signupSection.style.display = 'none';
 aboutSection.style.display = 'none';
 
+
+console.log(aboutSection.style.display + "awdadwa");
+
 const sections = {
     menu: menuSection,
     play: playSection,
@@ -34,7 +37,7 @@ const sections = {
 const openFunction = {
     play: startPlay(),
 
-    leaderBoard: openleaderBoard(),
+    //leaderBoard: openleaderBoard(),
 
     login: function () {
         const backButton = loginSection.getElementsByClassName('button-back')[0];
@@ -73,6 +76,7 @@ function openSection(section) {
     }
 
 }
+
 application.addEventListener('click', function (event) {
     const target = event.target;
 
@@ -102,86 +106,14 @@ function startPlay() {
     const leaderBoardButton = playSection.getElementsByClassName('button-leaderBoard')[0];
 
     backButton.addEventListener('click', getBack);
-    leaderBoardButton.addEventListener('click', openleaderBoard);
+    //leaderBoardButton.addEventListener('click', openleaderBoard);
 }
-function openleaderBoard(event) {
-    event.preventDefault();
+// function openleaderBoard(event) {
+//     event.preventDefault();
+//
+// }
 
 
-}
-
-class FormError {
-    constructor(formData, formType) {
-        this._data = formData;
-        this._type = formType;
-
-        this._stateError = {
-            usernameValid: false,
-            usernameError: '',
-
-            passwordValid: false,
-            passwordError: '',
-
-            passwordConfirmValid: false,
-            passwordConfirmError: ''
-        }
-    }
-
-    get stateError() {
-        return this._stateError;
-    }
-
-    validate() {
-        this.validateUsername();
-        this.validatePassword();
-
-        if (this._type === 'signup') {
-            this.validatePasswordConfirm();
-        }
-    }
-
-    validateUsername() {
-        if (!this._data['username']) {
-            this._stateError.usernameValid = false;
-            this._stateError.usernameError = 'Enter your username!\n';
-        } else {
-            this._stateError.usernameValid = true;
-            this._stateError.usernameError = '';
-        }
-    }
-
-    validatePassword() {
-        const regExp = /[^(\d\w)*]/;
-
-        if (!this._data['password']) {
-            this._stateError.passwordValid = false;
-            this._stateError.passwordError = 'Enter password!\n';
-
-        } else if (regExp.test(this._data['password'])) {
-            this._stateError.passwordValid = false;
-            this._stateError.passwordError = "Password must contain only latin symbols and digits\n";
-
-        } else {
-            this._stateError.passwordValid = true;
-            this._stateError.passwordError = '';
-        }
-    }
-
-    validatePasswordConfirm() {
-        if (!this._data['passwordConfirm']) {
-            this._stateError.passwordConfirmValid = false;
-            this._stateError.passwordConfirmError = 'Confirm password!\n';
-
-        } else if (this._data['password'] !== this._data['passwordConfirm']) {
-            this._stateError.passwordConfirmValid = false;
-            this._stateError.passwordConfirmError = 'Passwords do not match!\n';
-
-        } else {
-            this._stateError.passwordConfirmValid = true;
-            this._stateError.passwordConfirmError = '';
-        }
-    }
-}
 
 function getFormData(form, fields) {
